Refactor RegionMapping calculateImageryLayerIntervals() into a separate file.
